feat(IDENT-3086): Add Askar profile manager
