+++ requirements.askar.txt
@@ -1,3 +1,3 @@
-aries-askar~=0.2.0
+aries-askar~=0.2.2
 indy-credx~=0.3
 indy-vdr~=0.3
